chore(release): v1.0.0 (#91)
This release overwrites this library from a typescript (archived, deprecated) library to a swift one. Essentially, a direct pull of the natural language extensions from [Obsidian](https://github.com/beeauvin/obsidian). I'm moving on from typescript land unless it is required for something specific and don't have the energy to maintain that ecosystem anymore. To honor what folklore was, and the inspiration to create obsidian in the first place, these methods are now living on here. <3